src/utils/config.py
"""
Configuration loading utilities.
"""

import yaml
from pathlib import Path
from typing import Dict, Any
from omegaconf import OmegaConf


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load configuration from YAML file.
    
    Args:
        config_path: Path to config YAML file
    
    Returns:
        Configuration dictionary
    """
    config_path = Path(config_path)
    
    if not config_path.exists():
        raise FileNotFoundError(f"Config file not found: {config_path}")
    
    # Try OmegaConf first (supports variable interpolation)
    try:
        config = OmegaConf.load(config_path)
        return OmegaConf.to_container(config, resolve=True)
    except:
        # Fallback to PyYAML
        with open(config_path, "r") as f:
            config = yaml.safe_load(f)
        return config

