requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
scikit-learn>=1.3.0
pandas>=2.0.0
opencv-python>=4.7.0
einops>=0.6.0
accelerate>=0.20.0
clip-by-openai>=1.0
timm>=0.9.0
pyyaml>=6.0
omegaconf>=2.3.0

